Refactor comparison feature into smaller reusable modules.
Split picker, metrics, and shared actions into focused hooks and components, remove unused detail entity context, and lazy-load comparison data by variant.
