+++ src/components/explore/ComparisonAddButton.tsx
@@ -1,24 +1,16 @@
-import { useState } from "react";
 import { useTranslation } from "react-i18next";
 import { Plus } from "lucide-react";
 import { Button } from "@/components/ui/button";
 import { useComparison } from "@/contexts/ComparisonContext";
+import { getComparisonCopyKey } from "@/utils/explore/comparisonCopy";
 import type { ComparisonEntityVariant } from "@/utils/explore/comparisonUtils";
+import { useComparisonPickerTrigger } from "@/hooks/explore/useComparisonPickerTrigger";
 import { ComparisonPickerDialog } from "./ComparisonPickerDialog";
 
 interface ComparisonAddButtonProps {
   entityVariant?: ComparisonEntityVariant | null;
   variant?: "default" | "outline" | "ghost";
   size?: "default" | "sm" | "lg";
-}
-
-function addEntityLabelKey(
-  entityVariant: ComparisonEntityVariant | null | undefined,
-): string {
-  if (entityVariant) {
-    return `explorePage.comparison.addEntity.${entityVariant}`;
-  }
-  return "explorePage.comparison.addEntity.default";
 }
 
 export function ComparisonAddButton({
@@ -28,8 +20,10 @@ export function ComparisonAddButton({
 }: ComparisonAddButtonProps) {
   const { t } = useTranslation();
   const { variant: contextVariant } = useComparison();
-  const [open, setOpen] = useState(false);
   const entityVariant = entityVariantProp ?? contextVariant;
+  const { openPicker, dialogProps } = useComparisonPickerTrigger({
+    entityVariant,
+  });
 
   return (
     <>
@@ -37,16 +31,12 @@ export function ComparisonAddButton({
         variant={variant}
         size={size}
         className="gap-2"
-        onClick={() => setOpen(true)}
+        onClick={openPicker}
       >
         <Plus className="h-4 w-4" />
-        {t(addEntityLabelKey(entityVariant))}
+        {t(getComparisonCopyKey("addEntity", entityVariant))}
       </Button>
-      <ComparisonPickerDialog
-        open={open}
-        onOpenChange={setOpen}
-        entityVariant={entityVariant}
-      />
+      <ComparisonPickerDialog {...dialogProps} />
     </>
   );
 }
