+++ src/components/explore/comparisonPicker.constants.ts
@@ -1,0 +1,8 @@
+export const PICKER_RESULT_LIST_HEIGHT =
+  "h-36 max-h-36 min-h-36 [&_[cmdk-item]]:py-2.5";
+
+export const COMPARISON_SEARCH_CATEGORIES = [
+  "companies",
+  "municipalities",
+  "regions",
+] as const;
