+++ src/hooks/explore/useComparisonPickerTrigger.ts
@@ -1,0 +1,22 @@
+import { useState } from "react";
+import type { ComparisonPickerDialogProps } from "@/components/explore/comparisonPicker.types";
+
+type ComparisonPickerConfig = Omit<
+  ComparisonPickerDialogProps,
+  "open" | "onOpenChange"
+>;
+
+export function useComparisonPickerTrigger(config: ComparisonPickerConfig) {
+  const [open, setOpen] = useState(false);
+
+  return {
+    open,
+    setOpen,
+    openPicker: () => setOpen(true),
+    dialogProps: {
+      ...config,
+      open,
+      onOpenChange: setOpen,
+    } satisfies ComparisonPickerDialogProps,
+  };
+}
